+++ app/src/main/java/com/nextcloud/client/di/AppComponent.java
@@ -22,13 +22,13 @@ package com.nextcloud.client.di;
 
 import android.app.Application;
 
+import com.nextcloud.app_review.InAppReviewModule;
 import com.nextcloud.client.appinfo.AppInfoModule;
 import com.nextcloud.client.database.DatabaseModule;
 import com.nextcloud.client.device.DeviceModule;
 import com.nextcloud.client.integrations.IntegrationsModule;
 import com.nextcloud.client.jobs.JobsModule;
 import com.nextcloud.client.network.NetworkModule;
-import com.nmc.android.app_review.InAppReviewModule;
 import com.nextcloud.client.onboarding.OnboardingModule;
 import com.nextcloud.client.preferences.PreferencesModule;
 import com.owncloud.android.MainApp;
