--- app/src/main/java/com/nmc/android/app_review/AppReviewShownModel.kt
@@ -1,6 +1,0 @@
-package com.nmc.android.app_review
-
-data class AppReviewShownModel(var firstShowYear: String?,
-    var appRestartCount: Int,
-    var reviewShownCount: Int,
-    var lastReviewShownDate: String?)
