+++ app/src/main/java/com/owncloud/android/MainApp.java
@@ -40,7 +40,7 @@ import android.os.StrictMode;
 import android.text.TextUtils;
 import android.view.WindowManager;
 
-import com.nmc.android.app_review.InAppReviewHelper;
+import com.nextcloud.app_review.InAppReviewHelper;
 import com.nextcloud.client.account.User;
 import com.nextcloud.client.account.UserAccountManager;
 import com.nextcloud.client.appinfo.AppInfo;
