+++ app/src/main/java/com/owncloud/android/ui/activity/FileDisplayActivity.java
@@ -57,7 +57,7 @@ import android.view.WindowManager;
 
 import com.google.android.material.appbar.AppBarLayout;
 import com.google.android.material.snackbar.Snackbar;
-import com.nmc.android.app_review.InAppReviewHelper;
+import com.nextcloud.app_review.InAppReviewHelper;
 import com.nextcloud.client.account.User;
 import com.nextcloud.client.appinfo.AppInfo;
 import com.nextcloud.client.core.AsyncRunner;
