+++ app/src/qa/java/com/owncloud/android/app_review/InAppReviewHelperImpl.kt
@@ -23,8 +23,8 @@
 package com.owncloud.android.app_review
 
 import androidx.appcompat.app.AppCompatActivity
+import com.nextcloud.app_review.InAppReviewHelper
 import com.nextcloud.client.preferences.AppPreferences
-import com.nmc.android.app_review.InAppReviewHelper
 
 class InAppReviewHelperImpl(appPreferences: AppPreferences) :
     InAppReviewHelper {
